backend/requirements.txt
boto3>=1.34.0
flask>=3.0.0
python-dotenv>=1.0.0
flask-cors>=4.0.0
numpy>=1.26.0
opencv-python-headless>=4.9.0
Pillow>=10.0.0
